Move IAM credential fetching out of the s3 util.
All AWS services can use the IAM credentials, so it's ideal to split
it apart from the S3 utility.
